requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow
matplotlib
scikit-learn
split-folders
PyYAML
python-multipart
# -e .